Implement logging route and remove request logger middleware
