+++ agave/fastapi/middlewares/__init__.py
@@ -1,7 +1,5 @@
 from .error_handlers import AgaveErrorHandler
-from .request_logger import AgaveRequestLogger
 
 __all__ = [
     'AgaveErrorHandler',
-    'AgaveRequestLogger',
 ]
