+++ examples/fastapi/app.py
@@ -4,7 +4,7 @@ import mongomock as mongomock
 from fastapi import FastAPI
 from mongoengine import connect
 
-from agave.fastapi.middlewares import AgaveErrorHandler, AgaveRequestLogger
+from agave.fastapi.middlewares import AgaveErrorHandler
 
 from ..tasks.task_example import dummy_task, task_validator
 from .middlewares import AuthedMiddleware
@@ -20,7 +20,6 @@ app.include_router(resources)
 
 app.add_middleware(AuthedMiddleware)
 app.add_middleware(AgaveErrorHandler)
-app.add_middleware(AgaveRequestLogger)
 
 
 @app.get('/')
